Fixed Bug: Delete API is throwing 200 even though the id is not present. Added error handler and set to throw 404 exception

+++ src/main/java/com/karthik/StudentEntryApp/service/StudentsDepartmentImpl.java
@@ -5,6 +5,7 @@ import com.karthik.StudentEntryApp.entity.StudentsEntity;
 import com.karthik.StudentEntryApp.error.StudentIDNotFound;
 import com.karthik.StudentEntryApp.error.StudentNameNotFound;
 import com.karthik.StudentEntryApp.repository.StudentsRepository;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -13,6 +14,7 @@ import java.util.Objects;
 import java.util.Optional;
 import java.util.logging.Logger;
 
+@Slf4j
 @Service
 public class StudentsDepartmentImpl implements StudentsService{
     @Autowired
@@ -53,7 +55,11 @@ public class StudentsDepartmentImpl implements StudentsService{
     }
 
     @Override
-    public void deleteStudentById(Long id) {
+    public void deleteStudentById(Long id) throws StudentIDNotFound {
+        if(studentsRepository.findById(id).isEmpty()){
+            logger.info("No records found for the ID: " + id);
+            throw new StudentIDNotFound(String.format("Student ID: %s Not Found", id));
+        }
         studentsRepository.deleteById(id);
     }
